requirements.txt
robotframework
requests
robotframework-requests
robotframework-jsonlibrary
paramiko>=2.5.0
robotframework-sshlibrary
robotframework-scplibrary
redfish
click
PyYAML>=5.1
redfishtool
robotframework-selenium2library
webdrivermanager
robotframework-angularjs
robotframework-xvfb
redfish_utilities
sseclient
aenum
colorama
pyasn1
pyasn1_modules
robotframework-lint
